First commit of the repo.
As far as the testing were done, there are no compiling errors or other
issues with the repo or project. This project should work fine.
No bugs or errors were encountered by the author when publishing this
repo. So all the code in this project is working fine and is clean.
More details are provided in README.md