Improving Code and adding more methods to classes.
The code is added with a couple more methods in the Line class.
Also with a minor bug issue with accessing the static X_axis and
Y_axis being private. Now the bug is fixed by moving them to public.
These are minor changes added to the repository.

Date	: 13th July 2023
Time	: Arounf 22:00
Author	: Krishna Modepalli
Signature: krishnamodepalli